% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCout.R
\name{MCMCout}
\alias{MCMCout}
\title{Organize mcmc.list objects from coda/R2jags}
\usage{
MCMCout(mcmc.list)
}
\arguments{
\item{x}{an \code{mcmc.list} object.}

\item{...}{additional arguments passed to `density`, which 
determines posterior modes.}
}
\value{
a list containing arrays of 1) all posterior draws from all
  parallel chains, 2) posterior means, 3) posterior modes.
}
\description{
Get posterior draws, means, and modes output by MCMC simulations.
}
\details{
Takes \code{mcmc.list} objects, which are created by various 
functions in packages 'R2jags' and 'coda'. Returns posterior 
draws as arrays with dim(observations, output params, draws). Also
returns means and X values of modes (corresponding to highest posterior
peaks) as matrices with dim(observations, output params).
Works across all parallel MCMC chains (if running several in tandem).
}
\examples{
mod1 <- jags.model(model_script, data = list(<data>), 
        inits = function()list(<inits>), n.chains = 3, n.adapt = 2000)
update(mod1, n.iter=5e3)
mod1_out <- coda.samples(mod1, c(<variables out>, n.iter=1e5, thin=100))
MCMCout(mod1_out)
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\seealso{
\code{\link{MCMCout_mix}} for organized mixing model output.
}
\keyword{MCMC}
\keyword{R2jags}
\keyword{coda}

