% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCout_mix.R
\name{MCMCout_mix}
\alias{MCMCout_mix}
\title{Organize mcmc.list objects from coda/R2jags mixing models}
\usage{
MCMCout_mix(x, p, nsrc, ...)
}
\arguments{
\item{x}{an \code{mcmc.list} object from a mixing model.}

\item{...}{additional arguments passed to `density`, which 
determines posterior modes.}
}
\value{
a list containing arrays of 1) all posterior draws from all
  parallel chains, 2) posterior means, 3) posterior modes. Only 
  applies to the variable that contains source proportions.  For
  any other monitored variables, use \code{\link{MCMCout}}.
}
\description{
Get posterior draws, means, and modes for source proportions output by MCMC mixing model simulations.
}
\details{
Takes \code{mcmc.list} objects, which are created by various 
functions in packages 'R2jags' and 'coda'. Returns posterior 
draws as arrays with dim(sources, observed mixtures, draws). Also
returns means and X values of modes (corresponding to highest posterior
peaks) as matrices with dim(sources, observed mixtures).
Works across all parallel MCMC chains (if running several in tandem).
}
\examples{
mod1 <- jags.model(model_script, data = list(<data>), 
        inits = function()list(<inits>), n.chains = 3, n.adapt = 2000)
update(mod1, n.iter=5e3)
mod1_out <- coda.samples(mod1, c(<variables out>, n.iter=1e5, thin=100))
MCMCout_mix(mod1_out, 'prey_proportion', 3)
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\seealso{
\code{\link{MCMCout}} for organized output from non-mixing models.
}
\keyword{'mixing}
\keyword{MCMC}
\keyword{R2jags}
\keyword{coda}
\keyword{model'}

